src/test/java/uk/co/mruoc/wso2/DefaultUpdateApiUrlBuilderTest.java
package uk.co.mruoc.wso2;

import org.junit.Test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.Mockito.mock;

public class DefaultUpdateApiUrlBuilderTest {

    private static final String HOST_URL = "https://localhost:8443";
    private static final String RESOURCE = "/publisher/site/blocks/item-add/ajax/add.jag";
    private static final String QUERY_STRING = "?queryString=value";

    private final UpdateApiParamsToQueryStringConverter queryStringConverter = new StubUpdateApiParamsToQueryStringConverter(QUERY_STRING);
    private final UpdateApiUrlBuilder builder = new DefaultUpdateApiUrlBuilder(HOST_URL, queryStringConverter);

    @Test
    public void shouldBuildUrl() {
        AddApiParams params = mock(AddApiParams.class);

        String url = builder.build(params);

        assertThat(url).isEqualTo(HOST_URL + RESOURCE + QUERY_STRING);
    }

}
